サーバーサイドでのデータ取得が必要なコンポーネント構成の考察中
こんな感じで...Client.tsxと2つに分けてみるのはどうだろうか...
